refactor samples destroy_service into proejct and group specific services; fix associated tests
