+++ app/controllers/projects/samples/deletions_controller.rb
@@ -17,7 +17,7 @@ module Projects
       end
 
       def destroy
-        ::Samples::DestroyService.new(@project.namespace, current_user, { sample: @sample }).execute
+        ::Projects::Samples::DestroyService.new(@project, current_user, { sample: @sample }).execute
 
         respond_to do |format|
           if @sample.deleted?
@@ -34,11 +34,11 @@ module Projects
         end
       end
 
-      def destroy_multiple # rubocop:disable Metrics/AbcSize
+      def destroy_multiple
         samples_to_delete_count = destroy_multiple_params['sample_ids'].count
 
-        deleted_samples_count = ::Samples::DestroyService.new(@project.namespace, current_user,
-                                                              destroy_multiple_params).execute
+        deleted_samples_count = ::Projects::Samples::DestroyService.new(@project, current_user,
+                                                                        destroy_multiple_params).execute
 
         # No selected samples deleted
         if deleted_samples_count.zero?
