+++ test/components/projects/sample_activity_component_test.rb
@@ -68,7 +68,7 @@ module Projects
       ::Samples::UpdateService.new(sample, @user, valid_params).execute
 
       params = { sample: sample }
-      ::Samples::DestroyService.new(project_namespace, @user, params).execute
+      ::Projects::Samples::DestroyService.new(project_namespace.project, @user, params).execute
 
       activities = project_namespace.human_readable_activity(project_namespace.retrieve_project_activity).reverse
 
@@ -95,7 +95,7 @@ module Projects
       sample = samples(:sample1)
 
       params = { sample: sample }
-      ::Samples::DestroyService.new(project_namespace, @user, params).execute
+      ::Projects::Samples::DestroyService.new(project_namespace.project, @user, params).execute
 
       activities = project_namespace.human_readable_activity(project_namespace.retrieve_project_activity).reverse
 
@@ -123,7 +123,7 @@ module Projects
       sample2 = samples(:sample2)
 
       params = { sample_ids: [sample1.id, sample2.id] }
-      ::Samples::DestroyService.new(project_namespace, @user, params).execute
+      ::Projects::Samples::DestroyService.new(project_namespace.project, @user, params).execute
 
       activities = project_namespace.human_readable_activity(project_namespace.retrieve_project_activity).reverse
 
@@ -231,7 +231,7 @@ module Projects
 
       ::Attachments::DestroyService.new(sample, attachment, @user).execute
 
-      ::Samples::DestroyService.new(project_namespace, @user, { sample: sample }).execute
+      ::Projects::Samples::DestroyService.new(project_namespace.project, @user, { sample: sample }).execute
 
       activities = project_namespace.human_readable_activity(project_namespace.retrieve_project_activity).reverse
